Parameterize Tag based on the type it represents.
This is a nice little type-safety feature that can
catch small logic errors at compile time, e.g.
giving the tag to a valet of the wrong type.

It comes with a small cost of added code
complexity, since now we cannot use derives
because they place unnecessary bounds on the
type parameter.